requirements.txt
streamlit==1.40.0
streamlit-drawable-canvas==0.9.3
torch==2.6.0
torchvision==0.21.0
torch-fidelity==0.3.0
transformers==4.12.0
onnx==1.17.0
onnxruntime==1.21.0
timm==1.0.15
torchsummary
matplotlib
numpy
pandas
pillow
tqdm
python-telegram-bot